+++ database/seeders/CrVoucherTableSeeder.php
@@ -1,8 +1,10 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
-class DrVoucherTableSeeder extends Seeder
+class CrVoucherTableSeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -15,11 +17,11 @@ class DrVoucherTableSeeder extends Seeder
             'voucher_no' => '1',
             'branch_id'=>'1',
             'bank_cash_id'=>'1',
-            'income_expense_head_id'=>'1',
-            'voucher_type' => 'DV',
+            'income_expense_head_id'=>'8',
+            'voucher_type' => 'CV',
             'voucher_date'=>'2019-11-25',
-            'dr'=>'200',
-            'particulars'=>'Dr by System',
+            'cr'=>'100000',
+            'particulars'=>'CR by System',
             'created_by'=>'E-Accounts System',
         ]);
         App\Transaction::create([
@@ -27,36 +29,35 @@ class DrVoucherTableSeeder extends Seeder
             'branch_id'=>'2',
             'bank_cash_id'=>'2',
             'cheque_number'=>'112541',
-            'income_expense_head_id'=>'2',
-            'voucher_type' => 'DV',
+            'income_expense_head_id'=>'15',
+            'voucher_type' => 'CV',
             'voucher_date'=>'2019-11-26',
-            'dr'=>'500',
-            'particulars'=>'Dr by System',
+            'cr'=>'50000',
+            'particulars'=>'CR by System',
             'created_by'=>'E-Accounts System',
         ]);
         App\Transaction::create([
             'voucher_no' => '3',
-            'branch_id'=>'1',
+            'branch_id'=>'2',
             'bank_cash_id'=>'3',
             'cheque_number'=>'14521',
-            'income_expense_head_id'=>'3',
-            'voucher_type' => 'DV',
+            'income_expense_head_id'=>'15',
+            'voucher_type' => 'CV',
             'voucher_date'=>'2019-11-27',
-            'dr'=>'500',
-            'particulars'=>'Dr by System',
+            'cr'=>'15000',
+            'particulars'=>'CR by System',
             'created_by'=>'E-Accounts System',
         ]);
 
         App\Transaction::create([
             'voucher_no' => '3',
-            'branch_id'=>'1',
+            'branch_id'=>'2',
             'bank_cash_id'=>'3',
-            'cheque_number'=>'14521',
-            'income_expense_head_id'=>'4',
-            'voucher_type' => 'DV',
+            'income_expense_head_id'=>'16',
+            'voucher_type' => 'CV',
             'voucher_date'=>'2019-11-27',
-            'dr'=>'500',
-            'particulars'=>'Dr by System',
+            'cr'=>'60000',
+            'particulars'=>'Cr by System',
             'created_by'=>'E-Accounts System',
         ]);
     }
