+++ database/seeders/DrVoucherTableSeeder.php
@@ -1,8 +1,10 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
-class CrVoucherTableSeeder extends Seeder
+class DrVoucherTableSeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -15,11 +17,11 @@ class CrVoucherTableSeeder extends Seeder
             'voucher_no' => '1',
             'branch_id'=>'1',
             'bank_cash_id'=>'1',
-            'income_expense_head_id'=>'8',
-            'voucher_type' => 'CV',
+            'income_expense_head_id'=>'1',
+            'voucher_type' => 'DV',
             'voucher_date'=>'2019-11-25',
-            'cr'=>'100000',
-            'particulars'=>'CR by System',
+            'dr'=>'200',
+            'particulars'=>'Dr by System',
             'created_by'=>'E-Accounts System',
         ]);
         App\Transaction::create([
@@ -27,35 +29,36 @@ class CrVoucherTableSeeder extends Seeder
             'branch_id'=>'2',
             'bank_cash_id'=>'2',
             'cheque_number'=>'112541',
-            'income_expense_head_id'=>'15',
-            'voucher_type' => 'CV',
+            'income_expense_head_id'=>'2',
+            'voucher_type' => 'DV',
             'voucher_date'=>'2019-11-26',
-            'cr'=>'50000',
-            'particulars'=>'CR by System',
+            'dr'=>'500',
+            'particulars'=>'Dr by System',
             'created_by'=>'E-Accounts System',
         ]);
         App\Transaction::create([
             'voucher_no' => '3',
-            'branch_id'=>'2',
+            'branch_id'=>'1',
             'bank_cash_id'=>'3',
             'cheque_number'=>'14521',
-            'income_expense_head_id'=>'15',
-            'voucher_type' => 'CV',
+            'income_expense_head_id'=>'3',
+            'voucher_type' => 'DV',
             'voucher_date'=>'2019-11-27',
-            'cr'=>'15000',
-            'particulars'=>'CR by System',
+            'dr'=>'500',
+            'particulars'=>'Dr by System',
             'created_by'=>'E-Accounts System',
         ]);
 
         App\Transaction::create([
             'voucher_no' => '3',
-            'branch_id'=>'2',
+            'branch_id'=>'1',
             'bank_cash_id'=>'3',
-            'income_expense_head_id'=>'16',
-            'voucher_type' => 'CV',
+            'cheque_number'=>'14521',
+            'income_expense_head_id'=>'4',
+            'voucher_type' => 'DV',
             'voucher_date'=>'2019-11-27',
-            'cr'=>'60000',
-            'particulars'=>'Cr by System',
+            'dr'=>'500',
+            'particulars'=>'Dr by System',
             'created_by'=>'E-Accounts System',
         ]);
     }
