+++ database/seeders/IncomeExpenseHeadTableSeeder.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
 class IncomeExpenseHeadTableSeeder extends Seeder
